Distance field model

+++ CVPPP_Image2Label_dtfield.py
@@ -103,7 +103,8 @@ class Model(ModelDesc):
             
             with tf.name_scope('loss_mae'):
                 mae_il = tf.reduce_mean(tf.abs(pl - 
-                                             tf.cast(tf.cast(pil, tf.int32), tf.float32)), 
+                                             #tf.cast(tf.cast(pil, tf.int32), tf.float32)), 
+                                             pil),
                                         name='mae_il')
                 losses.append(1e0*mae_il)
                 add_moving_summary(mae_il)
